split app into different files
